Fix not found when eval as eval will not do dedup
Under eval mode, previously we will update the score.
It worked(found all existed including the duplicated) because we used a standalone lock API to update score, and find API will work for the duplicated keys.
But when we use find_and_update API to update the score, only 1 can
succeed for the duplicated keys, making the returned `founds` not
correct.